restore application dir and rename to papersfeed
