+++ backend/papers/models/keyword.py
@@ -5,3 +5,4 @@ from django.db import models
 class Keyword(models.Model):
     """Keyword Model"""
     name = models.CharField(max_length=200)
+
