+++ backend/papers/models/publication.py
@@ -14,3 +14,4 @@ class Publication(models.Model):
         null=False,
         on_delete=models.CASCADE,
     )
+
