+++ backend/papers/models/publisher.py
@@ -7,3 +7,4 @@ class Publisher(models.Model):
     name = models.CharField(max_length=200)
     city = models.CharField(max_length=20)
     address = models.CharField(max_length=100)
+
