+++ backend/papers/urls.py
@@ -6,8 +6,4 @@ from django.conf.urls import url
 from . import views
 
 urlpatterns = [
-    url(r'^(?P<api>\w+)$', views.api_entry, name='api_entry'),
-    url(r'^(?P<api>\w+)/(?P<second_api>\w+)$', views.api_entry, name='api_entry'),
-    url(r'^(?P<api>\w+)/(?P<second_api>\w+)/(?P<third_api>\w+)$', views.api_entry, name='api_entry'),
-    url(r'^(?P<api>\w+)/(?P<second_api>\w+)/(?P<third_api>\w+)/(?P<fourth_api>\w+)$', views.api_entry, name='api_entry')
 ]
