+++ backend/papers/utils.py
@@ -1,7 +1,2 @@
 """utils.py"""
 # -*- coding: utf-8 -*-
-
-
-class ApiError(Exception):
-    """raise ApiError"""
-    print('API ERROR(fixme)')
