+++ backend/papersfeed/admin.py
@@ -1,0 +1,4 @@
+"""admin.py"""
+#from django.contrib import admin
+
+# Register your models here.
