+++ backend/papersfeed/apps.py
@@ -1,0 +1,7 @@
+"""apps.py"""
+from django.apps import AppConfig
+
+
+class PapersfeedConfig(AppConfig):
+    """Papersfeed Config"""
+    name = 'papersfeed'
