+++ backend/papersfeed/models.py
@@ -1,0 +1,1 @@
+"""models.py"""
