+++ backend/papersfeed/urls.py
@@ -1,0 +1,13 @@
+"""urls.py"""
+# -*- coding: utf-8 -*-
+# pylint: skip-file
+
+from django.conf.urls import url
+from . import views
+
+urlpatterns = [
+    url(r'^(?P<api>\w+)$', views.api_entry, name='api_entry'),
+    url(r'^(?P<api>\w+)/(?P<second_api>\w+)$', views.api_entry, name='api_entry'),
+    url(r'^(?P<api>\w+)/(?P<second_api>\w+)/(?P<third_api>\w+)$', views.api_entry, name='api_entry'),
+    url(r'^(?P<api>\w+)/(?P<second_api>\w+)/(?P<third_api>\w+)/(?P<fourth_api>\w+)$', views.api_entry, name='api_entry')
+]
