+++ backend/papersfeed/utils.py
@@ -1,0 +1,7 @@
+"""utils.py"""
+# -*- coding: utf-8 -*-
+
+
+class ApiError(Exception):
+    """raise ApiError"""
+    print('API ERROR(fixme)')
